+++ packaging/postinstall.sh
@@ -12,6 +12,8 @@ cat > /etc/sudoers.d/micropanel <<EOF
 # Allow micropanel to run certbot and nginx without password
 micropanel ALL=(ALL) NOPASSWD: /usr/bin/certbot
 micropanel ALL=(ALL) NOPASSWD: /usr/sbin/nginx
+micropanel ALL=(ALL) NOPASSWD: /usr/bin/tee /etc/nginx/sites-enabled/*
+micropanel ALL=(ALL) NOPASSWD: /usr/bin/rm -f /etc/nginx/sites-enabled/*
 EOF
 chmod 440 /etc/sudoers.d/micropanel